Version 0.8
Added an arrow to better indicate cannon angle and initial speed and moved function get_screen_pos() out of the classes as it was very similar in all of then to a global function with the same name. Also angle and speed modifications of the cannon are handled with the call of some functions, witch update the velocity vector of the projectile at the same time.